Removed junk, added linux lib, platform detection
Removed some junk so files that don't need to be there.
Put the Windows .dll and new Linux .so in their own directory.
The code will now detect the type of platform it's running on and try to
load the appropriate library. Current supported platforms are Windows
and Linux. Other platforms will raise an exception.